backend/requirements.txt
fastapi==0.115.0
uvicorn==0.38.0
sqlalchemy==2.0.38
pydantic==2.10.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
pytest==8.3.4
httpx==0.28.1
asyncpg==0.30.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
email-validator==2.2.0
aiosqlite==0.21.0
pytest-asyncio==0.25.3